terminado

+++ src/main/java/JoseMerino_20240330/JoseMerino_20240330/Repository/LibrosRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LibrosRepository extends JpaRepository<LibrosEntity, Integer> {
+public interface LibrosRepository extends JpaRepository<LibrosEntity, Long> {
 }
